Removed dependency on ToolsSharp.
I decided to scrap the idea of a separate project with he core files
needed to write this lib. Instead, I created a "Core" folder and placed
all of the core files there. This means the core folder is now totally
independent of anything out side of it, and I am aiming to keep it that
way currently.